created validator for parkourpathway storage util

+++ src/main/java/org/braekpo1nt/mctmanager/games/game/parkourpathway/CheckPoint.java
@@ -3,5 +3,5 @@ package org.braekpo1nt.mctmanager.games.game.parkourpathway;
 import org.bukkit.Location;
 import org.bukkit.util.BoundingBox;
 
-public record CheckPoint(int yValue, BoundingBox boundingBox, Location respawn) {
+public record CheckPoint(double yValue, BoundingBox boundingBox, Location respawn) {
 }
